+++ src/sxw.js
@@ -1,22 +1,41 @@
 /**
 * @name         SXW.js
 * @description  A simple javascript to print self XSS warning
-* @gitHub       https://github.com/sxw-js/sxw-js
+* @gitHub       https://github.com/sxwjs/sxwjs
 * @website      https://sxw.js.org/
-* @version      v1.2.0 (2019-May-20)
+* @version      v2.0.0 (2019-May-20)
 * @license      MIT license (http://www.opensource.org/licenses/mit-license.php)
+* @author       Ref: https://github.com/sxwjs/sxwjs/graphs/contributors
 */
 
-(function (g) {
+/**
+* @description The template is based on https://github.com/umdjs/umd/blob/master/templates/returnExports.js
+*/
 
-    "/use strict";
+(function (root, factory) {
+    if (typeof define === 'function' && define.amd) {
+        // AMD.
+        // Register as an anonymous module.
+        define([], factory);
+    } else if (typeof module === 'object' && module.exports) {
+        // Node modules
+        // Does not work with strict CommonJS, but only CommonJS-like environments that support module.exports, like Node.
+        module.exports = factory();
+    } else {
+        // Browser globals (root is window)
+        root.sxwjs = factory();
+  }
+}(typeof self !== 'undefined' ? self : this, function () {
+
+    // Return a value to define the module export.
 
     var sxwjs = {
-        config: {
-            stopColor: "red",
-            warningText: "This section is intended for developers only. Don't copy paste anything in this area.\nIf someone told you to copy and paste something here, it is a scam and will give them access to your account. In that case, kindly report this to our support team."
-        }
     };
+
+    sxwjs.config = {
+        stopColor: "red",
+        warningText: "This section is intended for developers only. Don't copy paste anything in this area.\nIf someone told you to copy and paste something here, it is a scam and will give them access to your account. In that case, kindly report this to our support team."
+    }
 
     /**
      * @description This function prints a huge stop sign in the console
@@ -39,14 +58,11 @@
         printStop();
         printWarningText();
     }
-    
+
     // this assigning is done so that even when printStop and printWarningText are overridden printWarning remains unaffected
     sxwjs.printStop = printStop;
-    sxwjs.printWarningText = printWarningText;    
+    sxwjs.printWarningText = printWarningText;
     sxwjs.printWarning = printWarning;
 
-    // assinging sxwjs to window or global or the empty object that is passed
-    g.sxwjs = sxwjs;
-    
-})(window || global || {});
-// an empty object is passed in the absense of window or glbal to avoid errors
+    return sxwjs;
+}));